finished debugging. i can sleep now

+++ RationalNumber.java
@@ -18,7 +18,7 @@ public class RationalNumber extends RealNumber {
   }
 
   public double getValue() {
-    return numerator/denominator;
+    return (double)numerator/(double)denominator;
   }
 
   public int getNumerator() {
